add another password warning
I thought it would be a good idea to add another warning not to post the passwords inside config.php. The line "paste your config.php" still stands there as it is. This would look very bad on a screenshot in a blogpost how owncloud apparently didn't solve the problem. one more line doesn't hurt.